Added time validation for getting info from API.
Added time validation for retrieving currency rate information from API. From now on there is a mechanism to get the previous day's rates (during weekdays) until 12:15 (because the API data is updated on the NBP site until this time). After 12:15, the current day's rates are provided, as they are after the NBP API update.